Fix crypto include
Don't try to detect crypto availability in v3 and v5,
just require the module ourselves. If people try
to use it in-browser and have ignored the documentation,
they'll still get an error thrown, it just won't be as friendly.

A misunderstanding of the rng selection led to the previous code,
which worked in the REPL but not in any scripts.