build: set up and document build and release scripts
Since we have a build step now we must point the package.json file
reference to the dist folder and must explicitly list all files that we
want to be included in the npm tarball.

The initial idea was to prepare the contents of the final npm tarball
all in the dist directory (including a copy of the package.json) but
that turns out not to work particularly well with standard-version which
is used to manage the version numbers and changelog.

Unfortunately this change also required a small workaround to allow
local installation of the uuid module in the examples (see .local/).

Also instructions for relasing a new version have not yet been
documented which is finally done here.

Since we now explicitly list all files to be included in the tarball
through the files property of package.json we no longer need the
.npmignore file.