build: remove babel-plugin-add-module-exports
Since we have deprecated deep requires we no longer need the
module.exports from the individual v*.js files.

Since we now ship plain, non-babled deep require files on the top level
of the npm package for legacy CommonJS deep-require support we can
compensate for the lack of module.exports in there and instead simply
access the "default" export of the algorithm-specific modules.

This change was inspired by
https://github.com/webpack/webpack/issues/7973#issuecomment-417931938