fix: export package.json required by react-native and bundlers
It appears that react-native and some bundlers rely on being able to
introspect the package.json file of a npm module. After adding the
`exports` field to package.json, only the paths defined in that field
are accessible to Node.js. In order to support introspection of the
package.json file itself it has to be listed as an export as well.

See also:
https://github.com/ai/nanoevents/issues/44#issuecomment-602010343

Fixes #444