feat: remove deprecated v4 string parameter (#454)
In version 1.x of this library it was possible to call `v4('binary')` in
order to receive a byte array instead of a string representation. This
function signature was deprecated in 2.x (but not removed in 3.x as it
should have been).

The correct way to get a binary representation of a uuid is to pass an
array-like object as a second parameter:

```
const buffer = new Array();
v4(null, buffer);
```

Fixes #437.