test: install WebdriverIO packages as optional dev dependencies
WebdriverIO which we use for browser testing cannot be installed on some
platforms like z/OS.

In order to allow "Canary in the Goldmine"
(https://github.com/nodejs/citgm) to be run on such platforms as well we
need to install WebdriverIO as an optional devDependency. Unfortunately
npm doesn't support this out-of-the-box, so we have to wrap the affected
dependencies in a small local wrapper package.

Fixes #509.