build: drop support for legacy browsers (IE11, Safari 10)
BREAKING CHANGE: Drop support for browsers that don't correctly
implement const/let and default arguments, and no longer transpile the
browser build to ES2015.

This also removes the fallback on msCrypto instead of the crypto API.

Browser tests are run in the first supported version of each supported
browser and in the latest (as of this commit) version available on
Browserstack.